Adding location accuracy: Adjusted accuracy to ints, rounding edge case for accuracy

+++ networksurvey/src/main/java/com/craxiom/networksurvey/logging/SurveyRecordLogger.java
@@ -265,7 +265,7 @@ public abstract class SurveyRecordLogger
         tableColumns.add(FeatureColumn.createColumn(columnNumber++, MessageConstants.TIME_COLUMN, GeoPackageDataType.INT, false, null));
         tableColumns.add(FeatureColumn.createColumn(columnNumber++, MessageConstants.MISSION_ID_COLUMN, GeoPackageDataType.TEXT, false, null));
         tableColumns.add(FeatureColumn.createColumn(columnNumber++, MessageConstants.RECORD_NUMBER_COLUMN, GeoPackageDataType.MEDIUMINT, true, -1));
-        tableColumns.add(FeatureColumn.createColumn(columnNumber++, MessageConstants.ACCURACY, GeoPackageDataType.FLOAT, false, null));
+        tableColumns.add(FeatureColumn.createColumn(columnNumber++, MessageConstants.ACCURACY, GeoPackageDataType.INT, false, null));
 
         if (addCellularColumns)
         {
